create config files in the user's home dir from a zip template
This new feature reads a list of config files from a zip file named
config/configtemplates.zip and checks each file, if it is already
existing in the game's search path. If it can not be found, it will
be created in the user's home dir (where saved.cfg gets written).

Before this, the config files were installed with the rest of the game,
and if the user customised any of the files, he most likely did not move
the file to his personal directory first. A re-install in the same place
would then overwrite the customised file.

This feature creates the files where they will usually not be overwritten.

To use the feature, before packaging the game, the files
 config/autoexec.cfg
 config/favourites.cfg
 config/pcksources.cfg
have to be archived into a zip file
 config/configtemplates.zip

The zip file should not contain directories. It should not be checked
into git. The original single config files should not be distributed
with the package.