removed huge unnecessary file
The file packages/models/playermodels/RVSF/07_Blue-Red.psd had a filesize
of over nine megabyte - and was of no use to the game. It made all recent
game packages 9MB bigger than they had to be. That's about the size of
all game textures combined...
If someone still needs the file, check it out from an older commit.