Implement the INSTA CTF mode
The server is made to automatically respawn players without delay when
the gamemode is set to INSTACTF. The client will not display a timer,
and allow the request for respawn to be sent to the server.

The macros now use the appropriate GMODE_ enum, instead of inlining
the values directly.